Can get alternatives to print to console
