+++ modules/users/client/controllers/recipes/addRecipe.client.controller.js
@@ -10,41 +10,111 @@
   function AddRecipeController(UsersService, $scope, $http) {
     var vm = this;
 
-    var alternative = {
-      'hello' : 'world'
+    // var alternative = {
+    //   'hello' : 'world'
+    // }
+
+    // API KEY
+		var apiKey = 'YAJ2M9l67OaqNMPCEfBcoccVtQDY5LPUR20rFzP8';
+
+		// FOR REPORT
+		var type = "b";
+		var format = "json";
+
+		// FOR INDIVIDUAL SEARCHES
+		var sort = "n";
+		var max = "200";
+		var ds = 'Standard Reference';
+
+    $scope.map = [];
+		$scope.in_food_group;
+		$scope.orig_nutrient_amount;
+		$scope.all_alt_in_group = [];
+		$scope.have_match = 0;
+		$scope.c_method = "baked";
+    $scope.search = "butter";
+
+    function getAlternatives() {
+      $http.get('./modules/users/client/controllers/recipes/food_alternatives.json')
+        .then( (response) => {
+          response.data.cooking_methods.forEach( (cooking_method, i) => {
+						cooking_method.food_groups.forEach( (food_group, j) => {
+							food_group.food_alts.forEach( (food_alt, k) => {
+
+								if((food_alt.db_name == $scope.search) && ($scope.c_method == cooking_method.method_name)){
+									$scope.have_match = 1;
+									$scope.orig_ndbno = food_alt.db_ndbno;
+									$scope.in_food_group = food_group.group_name;
+									$scope.orig_nutrient_amount = food_alt.db_main_nutrient.db_amount;
+								}
+								else{
+									$scope.all_alt_in_group.push(food_alt);
+								}
+							});
+							if($scope.have_match == 1){
+								$scope.all_alt_in_group.forEach((alt_item, i) => {
+									if(alt_item.db_main_nutrient.db_amount < $scope.orig_nutrient_amount){
+										$scope.map.push({"map_ndbno": alt_item.db_ndbno, "map_name": alt_item.db_name});
+									}
+								});
+							}
+							$scope.have_match = 0;
+							$scope.all_alt_in_group = [];
+						});
+					});
+        });
     }
 
-    $http.get('food_alternatives.json')
-      .then( (response) => {
-        console.log(response);
-      });
+    $scope.getReport = (searchedItem) => {
+			var reportURL = 
+			  	"http://api.nal.usda.gov/ndb/reports/" + 
+			  	"?ndbno=" + searchedItem + 
+          "&type=" + type + 
+          "&format=" + format + 
+          "&api_key=" + apiKey; 
 
-    UsersService.usdaAlternatives(alternative)
-      .then(success)
-      .catch(failure)
+        getURL(reportURL)
+          .then( (results) => {
+            $scope.searched = results.data;
+            assignFood();
+          });
+		}
 
-    function success(response) {
-      console.log('worked!');
-      console.log(response);
-    }
+    function assignFood() {
+			$scope.food = $scope.searched.report.food.name.toLowerCase();
+			// $scope.ingredients = $scope.searched.report.food.ing.desc.toLowerCase();
+			$scope.nutrients = $scope.searched.report.food.nutrients;
+		}
 
-    function failure(response) {
-      console.log('sadness')
-      console.log(response);
-    }
+    function getURL(url) {
+			return $http.get(url);
+		}
+
+    getAlternatives();
+    console.log($scope.map);
+
+
+    // UsersService.usdaAlternatives(alternative)
+    //   .then(success)
+    //   .catch(failure)
+
+    // function success(response) {
+    //   console.log('worked!');
+    //   console.log(response);
+    // }
+
+    // function failure(response) {
+    //   console.log('sadness')
+    //   console.log(response);
+    // }
 
     //DO YOUR FRONTEND JS CODE HERE
-//===========================================================================
     $scope.recipeList = [{}];
 
     $scope.recipeAdd = function() {
         $scope.recipeList.push({});
     };
-//===========================================================================
 
-    // $scope.alert = () => {
-  	// 	alert('hello');
-  	// }
 
     //vm.addings = UsersService.query();
   }
